walkfind-ml-worker/requirements.txt
boto3==1.35.0
psycopg2-binary==2.9.9
pgvector==0.3.6
python-dotenv==1.0.1
Pillow==10.4.0

torch==2.3.1
open-clip-torch==2.26.1